ShouldExportRecordAsync => FAIL
