+++ RedcapApi/Services/IApiService.cs
@@ -1,0 +1,9 @@
+﻿using System.Threading.Tasks;
+
+namespace Redcap.Services
+{
+    public interface IApiService
+    {
+        Task<T> ExportRecordAsync<T>();
+    }
+}
